config: Add support for default options
Support allowing a user to specify an option to override the chosen default.
For example, the user can set `DefaultDistribution=` and the configuration
will automatically resolve `state.config.distribution`. The user can specify
`Distribution=` in a preset or configuration file and override the default.

Default values are only allowed to be set in configuration files.