Don't run depmod unless needed
For distro kernels, depmod is already executed by the package
manager, so let's make sure we don't rerun it unless needed.

To make this work, we have to run depmod and modinfo from inside the
image, as running modinfo from the host on files generated by depmod
from the image can lead to all sorts of compatibility issues so we opt
to run both depmod and modinfo from inside the image.