Make sure we always clean up /work
Currently we only clean up /work if the build script succeeds.
However, we have to always clean up /work since we don't make a
copy of the build overlay so let's make sure we do that by using
a trap instead of only cleaning up if a script succeeds.

Let's also make sure that we don't fail if /work already exists and
make sure that if gets the right permissions, even if it already
exists.