initrd: add mkosi.finalize that masks initrd-udevadm-cleanup-db.service
Without this, machines with lvm2 do not boot. This code is copied from
mkosi-initrd and allows the initrds to work with non-trivial storage.

I expect that there will be other modifications to make, so the script is
written in python in a way that should make it easy to extend in the future.