mkosi-initrd: Only include core + host's loaded modules and firmware
The initramfs produced by mkosi-initrd's kernel-install plugin is only
intended to be used on the host system that's building it. Let's take
advantage of that and only include the currently loaded modules and
firmware in the initramfs to cut down the size.

For safety reasons we always include some core modules regardless of
whether they're loaded on the host or not.